# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: takos API
  version: "1.0.0"
  description: |
    takos のバックエンド API を OpenAPI 形式で記述した仕様書です。
    認証はセッション Cookie を利用します。
servers:
  - url: https://{domain}
    description: 本番環境
    variables:
      domain:
        default: example.com
  - url: http://localhost:8000
    description: ローカル開発環境
paths:
...
  /api/dms:
    get:
      summary: DM一覧取得
      parameters:
        - name: owner
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: DMの配列
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    owner:
                      type: string
                    name:
                      type: string
                    icon:
                      type: string
                    members:
                      type: array
                      items:
                        type: string
    post:
      summary: DM作成
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                owner:
                  type: string
                id:
                  type: string
                name:
                  type: string
                members:
                  type: array
                  items:
                    type: string
              required:
                - owner
                - id
                - name
      responses:
        "200":
          description: 作成されたDM
  /api/dms/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    patch:
      summary: DM更新
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                owner:
                  type: string
                name:
                  type: string
              required:
                - owner
      responses:
        "200":
          description: 更新後のDM
    delete:
      summary: DM削除
      parameters:
        - name: owner
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 削除結果
  /api/dms/{id}/icon:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: owner
        in: query
        required: true
        schema:
          type: string
    post:
      summary: DMアイコンアップロード
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
                - file
      responses:
        "200":
          description: アップロードされたアイコンのURL
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
...